Fix nits and structure
