+++ src/main/webapp/planner.jsp
@@ -1,4 +1,4 @@
-<jsp:include page="template.jsp">
+<jsp:include page="vendor/startbootstrap/template.jsp">
   <jsp:param name="content" value="planner-content"/>
   <jsp:param name="scriptFile" value="plannerScript"/>
 </jsp:include>
